% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias_adjustment.R
\name{de_grid_one_pi_multi_kappa}
\alias{de_grid_one_pi_multi_kappa}
\title{Estimate bias-corrected DE for a single pi_param and a grid of kappa values}
\usage{
de_grid_one_pi_multi_kappa(
  Y_e,
  Z_e,
  mu_e_1,
  mu_e_0,
  pi_vec,
  kappa_vec_,
  pz,
  estimate_var = FALSE
)
}
\arguments{
\item{Y_e}{Vector of observed ego outcomes.}

\item{Z_e}{Vector of observed ego treatments.}

\item{mu_e_1}{Vector of predictions E[Y|Z=1, X].}

\item{mu_e_0}{Vector of predictions E[Y|Z=0, X].}

\item{pi_vec}{Vector of pi_i^e values.}

\item{kappa_vec_}{Vector of kappa values to iterate over.}

\item{pz}{Scalar Pr(Z=1).}

\item{estimate_var}{Logical. If TRUE, compute empirical variance.}
}
\value{
A data.table with columns: kappa, de_rd, de_rd_var
}
\description{
Estimate bias-corrected DE for a single pi_param and a grid of kappa values
}
\keyword{internal}
