% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias_adjustment.R
\name{de_point_one_pi_kappa}
\alias{de_point_one_pi_kappa}
\title{Calculate bias-corrected DE_RD for a single (pi, kappa) combination}
\usage{
de_point_one_pi_kappa(
  Y_e,
  Z_e,
  mu_e_1,
  mu_e_0,
  pi_vec,
  kappa_,
  pz,
  estimate_var = FALSE
)
}
\arguments{
\item{Y_e}{Vector of observed ego outcomes.}

\item{Z_e}{Vector of observed ego treatments.}

\item{mu_e_1}{Vector of predictions m^e_i(1) = E[Y|Z=1, X].}

\item{mu_e_0}{Vector of predictions m^e_i(0) = E[Y|Z=0, X].}

\item{pi_vec}{Vector of pi_i^e values.}

\item{kappa_}{Scalar kappa value.}

\item{pz}{Scalar Pr(Z=1).}

\item{estimate_var}{Logical. If TRUE, compute empirical variance.}
}
\value{
A numeric vector `c(de_rd, de_rd_var)`
}
\description{
Implements the augmented randomization-based estimator:
DE_aug = (1/n_e) * sum[ w_i * phi_i ]
where w_i = 1 / (1 + pi_i * (kappa - 1))
and   phi_i = [ I(Z=1)(Y-m1)/pz - I(Z=0)(Y-m0)/(1-pz) + (m1 - m0) ]
}
\keyword{internal}
