% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ENRT_sensitivity-package.R
\docType{package}
\name{ENRTsensitivity-package}
\alias{ENRTsensitivity}
\alias{ENRTsensitivity-package}
\title{ENRTsensitivity: Sensitivity Analysis for Contamination in Egocentric-Network Randomized Trials}
\description{
Implements the sensitivity analysis framework for egocentric-network randomized trials (ENRTs) proposed by Weinstein and Nevo (2025). Provides tools to conduct grid sensitivity analysis (GSA) and probabilistic bias analysis (PBA) to assess the impact of ego-network contamination on indirect and direct effect estimates.
}
\author{
\strong{Maintainer}: Bar Weinstein \email{weinstein.bar@gmail.com}

}
\keyword{internal}
