% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias_adjustment.R
\name{ie_aug_point_grid_}
\alias{ie_aug_point_grid_}
\title{Calculate augmented IE estimates for a grid of pi_params}
\usage{
ie_aug_point_grid_(
  Y_a,
  F_a,
  mu_a_1 = NULL,
  mu_a_0 = NULL,
  pi_list,
  pz,
  ego_id_a = NULL,
  n_e = NULL,
  folds_ids_a = NULL
)
}
\arguments{
\item{Y_a}{Vector of alter outcomes.}

\item{F_a}{Vector of alter observed exposures.}

\item{mu_a_1}{Vector of cross-fit predictions E[Y|F=1,X].}

\item{mu_a_0}{Vector of cross-fit predictions E[Y|F=0,X].}

\item{pi_list}{List of pi_vec sensitivity parameters.}

\item{pz}{Scalar, Pr(Z=1) among egos.}

\item{ego_id_a}{Vector mapping alters to ego indices (1 to n_e).}

\item{n_e}{Total number of egos.}

\item{folds_ids_a}{Vector indicating fold assignments for alters.}
}
\value{
A data.table with pi_param, ie_rd, ie_rd_var
}
\description{
Calculate augmented IE estimates for a grid of pi_params
}
\keyword{internal}
