% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias_adjustment.R
\name{de_grid_multi_pi_kappa}
\alias{de_grid_multi_pi_kappa}
\title{Estimate bias-corrected DE estimates for a grid of (pi_param, kappa) combinations}
\usage{
de_grid_multi_pi_kappa(
  Y_e,
  Z_e,
  mu_e_1 = NULL,
  mu_e_0 = NULL,
  pi_list,
  kappa_vec,
  pz,
  estimate_var = FALSE
)
}
\arguments{
\item{Y_e}{Vector of observed ego outcomes (n_e).}

\item{Z_e}{Vector of observed ego treatments (n_e).}

\item{mu_e_1}{Vector of predictions E[Y|Z=1, X] (n_e).}

\item{mu_e_0}{Vector of predictions E[Y|Z=0, X] (n_e).}

\item{pi_list}{A list of pi_vecs.}

\item{kappa_vec}{A vector of kappa values.}

\item{pz}{Scalar Pr(Z=1).}

\item{estimate_var}{Logical. If TRUE, compute empirical variance.}
}
\value{
A data.table with columns: pi_param, kappa, de_rd, de_rd_var.
}
\description{
Estimate bias-corrected DE estimates for a grid of (pi_param, kappa) combinations
}
\keyword{internal}
